requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.20.0
scikit-learn>=1.4.0
xgboost>=2.0.3
scipy>=1.12.0
psutil>=5.9.8
pyyaml>=6.0.1
joblib>=1.3.2
fastapi>=0.115.0
uvicorn>=0.30.6
